"C-x 2"/"C-x 3": Do helm-mini instead of switching to previous.
Sometimes splitting and showing previous buffer will have a different directory
than the current buffer, so it is not expected behavior when just wanting to
view another file of the same directory. This way it's easy to choose what the
buffer should display or to find another file starting from the same directory.