Auto-compiling confs, no recompile on close.
After loading all packages asynchronously it checks if the .elc files are not there, if not then it
compiles general.el, functions.el, and all init.d/*.el files. This is important because if the .elc
files are not there then saving an .el file won't create the corresponding .elc file.