Auto-create dirs after `find-file` and saving buffer
If not saving buffer and killing it, it will ask to delete all the created dirs going upwards.

Also works with `find-alternate-file` and `write-file`.

Automatically create and delete parent directories of new files. This advice automatically creates
the parent directory (or directories) of the file being visited, if necessary. It also sets a
buffer-local variable so that the user will be prompted to delete the newly created directories if
they kill the buffer without saving it.

This advice has no effect for remote files.

This is an `:around' advice for `find-file' and similar functions.

FIND-FILE is the original `find-file'; FILENAME and ARGS are its arguments.

Originally found at:
https://github.com/radian-software/radian/blob/a8bb096c80d2daf34d380331fb23fa338ac1bb17/emacs/radian.el#L1112